requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
dj-rest-auth==2.2.4
Django==4.0.5
django-debug-toolbar==3.5.0
django-heroku==0.3.1
django-super-inlines==0.2.1
django-utils-six==2.0
djangorestframework==3.13.1
djangorestframework-filters==0.11.1
drf-yasg==1.20.0
gunicorn==20.1.0
heroku==0.1.4
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
psycopg2==2.9.3
pyparsing==3.0.9
python-dateutil==1.5
python-decouple==3.6
pytz==2022.1
requests==2.28.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.2.0
